Split flashcard.py into deck.py and flashcard.py per repo conventions
